+++ Tests/Tests.Actors.BugFinding/Integration/ReplicatingStorageTests.cs
@@ -856,7 +856,7 @@ namespace Microsoft.Coyote.Actors.BugFinding.Tests
 
         [Theory(Timeout = 10000)]
         [InlineData(1)]
-        public void TestReplicatingStorageLivenessBug(ulong seed)
+        public void TestReplicatingStorageLivenessBug(uint seed)
         {
             var configuration = this.GetConfiguration();
             configuration.MaxUnfairSchedulingSteps = 200;
